@@ -10,17 +10,18 @@
   },
   "dependencies": {
     "lodash-es": "^4.17.21",
-    "rxjs": "^6.5.5",
+    "rxjs": "^6.6.7",
     "tslib": "^2.3.1",
-    "vue": "^2.6.11",
+    "vue": "^2.6.14",
     "vue-class-component": "^7.2.6",
-    "vue-property-decorator": "^8.4.1"
+    "vue-property-decorator": "^8.5.1"
   },
   "devDependencies": {
     "@types/chrome": "0.0.99",
     "@types/firefox-webext-browser": "^67.0.2",
     "@types/lodash": "^4.14.178",
-    "@types/lodash-es": "^4.17.5",
+    "@types/lodash-es": "^4.17.6",
+    "@types/uuid": "^8.3.4",
     "@typescript-eslint/eslint-plugin": "^2.26.0",
     "@typescript-eslint/parser": "^2.26.0",
     "@vue/cli-plugin-eslint": "~4.2.0",
@@ -29,14 +30,14 @@
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^5.0.1",
     "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-prettier": "^3.4.1",
     "eslint-plugin-vue": "^6.2.2",
-    "git-revision-webpack-plugin": "^3.0.4",
-    "less": "^3.0.4",
+    "git-revision-webpack-plugin": "^3.0.6",
+    "less": "^3.13.1",
     "less-loader": "^5.0.0",
-    "prettier": "^2.0.2",
-    "typescript": "^3.8.3",
+    "prettier": "^2.5.1",
+    "typescript": "^3.9.10",
     "vue-cli-plugin-browser-extension": "~0.24.0",
-    "vue-template-compiler": "^2.6.11"
+    "vue-template-compiler": "^2.6.14"
   }
 }
