@@ -1,20 +1,26 @@
-import { GetUsersResponse } from '@/types/user.response';
-import { StreamByUserResponse, LiveStreamsResponse } from '@/types/stream.response';
-import { ChannelResponse } from '@/types/channel.response';
-import { FeaturedStreamsResponse } from '@/types/featured.response';
-import { TopGamesResponse, SearchGamesResponse } from '@/types/categories.response';
+import { GetUsersResponse } from '@/types/users.response';
+import { GetStreamsResponse } from '@/types/streams.response';
+import { Channel, GetChannelInformationResponse } from '@/types/channel.response';
+import { TopCategoriesResponse, Category, SearchCategoriesResponse } from '@/types/categories.response';
+import { TokenManager } from './token-manager';
 
 export class TwitchAPI {
-  private static endpoint = 'https://api.twitch.tv/kraken';
-  private static clientId = 'sf954451u6y5bcsilwklkw4tsd8za3';
-
-  private static async request<T>(path: string): Promise<T> {
-    const response = await fetch(this.endpoint + path, {
+  private static async request<T>(path: string, retried = false): Promise<T> {
+    const token = await TokenManager.getAccessToken();
+    const response = await fetch(`https://api.twitch.tv/helix${path}`, {
       headers: {
-        'Client-ID': this.clientId,
-        accept: 'application/vnd.twitchtv.v5+json',
+        Authorization: `Bearer ${token}`,
+        'Client-Id': process.env.VUE_APP_CLIENT_ID,
       },
     });
+
+    if (response.status === 401) {
+      await TokenManager.deleteTokens();
+      if (!retried) {
+        return await this.request(path, true);
+      }
+    }
+
     if (response.status !== 200) {
       throw new Error('Bad response from server');
     }
@@ -22,42 +28,43 @@ export class TwitchAPI {
     return response.json();
   }
 
-  public static async getChannelId(name: string): Promise<string | null> {
-    const response = await this.request<GetUsersResponse>(`/users/?login=${name}`);
-    return response._total === 0 ? null : response.users[0]._id;
+  public static async getChannelId(username: string): Promise<string | null> {
+    const response = await this.request<GetUsersResponse>(`/users/?login=${username}`);
+    return response.data[0]?.id ?? null;
   }
 
   public static async isOffline(id: string): Promise<boolean> {
     const response = await this.getStreamsByIds(id);
-    return response.streams.length === 0;
+    return response.data.filter((x) => x.type === 'live').length === 0;
   }
 
-  public static async getChannelById(id: string) {
-    return this.request<ChannelResponse>(`/channels/${id}`);
+  public static async getChannelById(id: string): Promise<Channel | null> {
+    const response = await this.request<GetChannelInformationResponse>(`/channels/?broadcaster_id=${id}`);
+    return response.data[0] ?? null;
   }
 
-  public static async getStreamsByIds(ids: string | string[]): Promise<LiveStreamsResponse> {
+  public static async getStreamsByIds(ids: string | string[]): Promise<GetStreamsResponse> {
     const idsArray = Array.isArray(ids) ? ids : [ids];
-    return this.request<LiveStreamsResponse>(`/streams/?channel=${idsArray.join(',')}`);
+    return this.request<GetStreamsResponse>(`/streams/?user_id=${idsArray.join(',')}`);
   }
 
   public static async getStreamsByCategory(category: string) {
-    const response = await this.request<LiveStreamsResponse>(`/streams/?game=${encodeURIComponent(category)}`);
-    return response.streams;
+    const response = await this.request<GetStreamsResponse>(`/streams/?game_id=${category}`);
+    return response.data;
   }
 
-  public static async getFeaturedStreams() {
-    const response = await this.request<FeaturedStreamsResponse>('/streams/featured');
-    return response.featured;
+  public static async getPopularStreams() {
+    const response = await this.request<GetStreamsResponse>('/streams');
+    return response.data;
   }
 
-  public static async getTopCategories() {
-    const response = await this.request<TopGamesResponse>('/games/top?limit=30');
-    return response.top.map((x) => x.game);
+  public static async getTopCategories(): Promise<Category[]> {
+    const response = await this.request<TopCategoriesResponse>('/games/top?first=30');
+    return response.data;
   }
 
-  public static async searchCategories(query: string) {
-    const response = await this.request<SearchGamesResponse>(`/search/games?type=suggest&live=true&query=${query}`);
-    return response.games || [];
+  public static async searchCategories(query: string): Promise<Category[]> {
+    const response = await this.request<SearchCategoriesResponse>(`/search/categories?query=${query}`);
+    return response.data;
   }
 }
